Update simplybook webhook
